BugFix - uint8_t in fftPeaksInfor, refactoring

+++ App/inc/radio_protocol.h
@@ -56,7 +56,7 @@ typedef struct
 	float32_t rms;
 	float32_t temperature;
 	uint8_t   fftPeaksNum;
-	uint16_t  fftPeaksInfo[];
+	uint8_t  fftPeaksInfo[];
 
 } __PACKED__ radioprot_status_info_t;
